requirements.txt
fastapi
google-genai
langchain
langchain-core
langchain-google-genai
langgraph
langsmith
pydantic
python-dotenv
uvicorn
